一週間以前のログを対象にS3に転送してログを削除する

+++ cloudwatchlogToS3/main.py
@@ -1,5 +1,6 @@
 import boto3
 import datetime
+from _datetime import date, timedelta
 
 # ロググループを検索する際のクエリ―
 PREFIX = '/aws/lambda/hmkvDementiaFunction'
@@ -37,10 +38,9 @@ def lambda_handler(event=None, context=None):
         )
         log_storeams.extend(log_stoream_response['logStreams'])
 
-    # 実行日時の0:00:00~23:59:59を取得
+    # 一週間前のログを対象とする
     now = datetime.datetime.now()
-    start_day = datetime.datetime(now.year, now.month, now.day, 0, 0, 0)
-    end_day = datetime.datetime(now.year, now.month, now.day, 23, 59, 59)
+    end_day = now - timedelta(days=7)
 
     # エクスポートタスクを作成
     response = client.create_export_task(
@@ -54,10 +54,11 @@ def lambda_handler(event=None, context=None):
     if response['ResponseMetadata']['HTTPStatusCode'] == 200:
         # 転送したログストリームを削除する
         for log_stoream in log_storeams:
-            client.delete_log_stream(
-                logGroupName=log_group['logGroupName'],
-                logStreamName=log_stoream['logStreamName']
-            )
+            if log_stoream['creationTime'] < int(end_day.timestamp() * 1000):
+                client.delete_log_stream(
+                    logGroupName=log_group['logGroupName'],
+                    logStreamName=log_stoream['logStreamName']
+                )
 
     return response
